Rearrange and expand content-type checks in hybrid.spec tests
The changes in this commit reflect a rearrangement of the expectation order in hybrid.spec.ts and expansion of the 'content-type' header checks. To better reflect the order of operations, 'preRender.status' check has been moved. Besides, additional checks have been introduced for 'content-type' headers for 'indexRender' and 'echoReply'. The check for 'preRender' is commented out as it doesn't function as expected due to an existing issue.